requirements.txt
httptools
Jinja2
matplotlib
pandas
pydantic-settings
python-dotenv
python-multipart
requests==2.31.0
redis
aiokafka
asyncpg
fastapi
greenlet
prometheus_client
psycopg2-binary
sqlalchemy[asyncpg]
uhashring
uvicorn
uvloop
